+++ pragma-ui/src/app/assets/_helpers/getTokens.ts
@@ -1,5 +1,5 @@
-import { dataSources } from "@/lib/endpoints";
 import { Token, AssetT } from "@/app/assets/_types";
+import { DEFAULT_ASSETS } from "@/lib/constants";
 
 export const getTokens = async (source?: string): Promise<AssetT[]> => {
   if (source === "api") {
@@ -16,22 +16,5 @@ export const getTokens = async (source?: string): Promise<AssetT[]> => {
     return data.tokens || [];
   }
 
-  return [
-    { ticker: "BTC/USD", address: "0x2", decimals: 8 },
-    { ticker: "ETH/USD", address: "0x1", decimals: 8 },
-    { ticker: "WBTC/USD", address: "0x3", decimals: 8 },
-    { ticker: "STRK/USD", address: "0x4", decimals: 8 },
-    { ticker: "WSTETH/USD", address: "0x5", decimals: 8 },
-    { ticker: "LORDS/USD", address: "0x6", decimals: 8 },
-    { ticker: "EKUBO/USD", address: "0x7", decimals: 8 },
-    { ticker: "BROTHER/USDPLUS", address: "0x8", decimals: 8 },
-    { ticker: "ZEND/USD", address: "0x9", decimals: 8 },
-    { ticker: "DAI/USD", address: "0x9", decimals: 8 },
-    { ticker: "USDC/USD", address: "0x9", decimals: 8 },
-    { ticker: "USDT/USD", address: "0x9", decimals: 8 },
-    { ticker: "BTC/EUR", address: "0x9", decimals: 8 },
-    { ticker: "WBTC/BTC", address: "0x9", decimals: 8 },
-    { ticker: "NSTR/USD", address: "0x9", decimals: 8 },
-    { ticker: "XSTRK/USD", address: "0x9", decimals: 8 },
-  ];
+  return DEFAULT_ASSETS;
 };
